pyproject.toml
[project]
name = "zoom-pro"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "duckdb>=1.5.1",
    "google-cloud-storage>=3.10.0",
    "kagglehub[pandas-datasets]>=1.0.0",
    "notebook>=7.5.5",
]
